Extract UploadResponseController
This should be its own thing. More work to do to split it out into
separate actions for GET & POST, improve the routes, lots of code
cleanup, etc, but at least now it's in its own file to start that
process.
